requirements.txt
tweepy
praw
requests
langchain
chromadb
sentence-transformers
python-dotenv
pandas
numpy
matplotlib
scikit-learn
nltk
transformers
torch
ragas
tqdm